*Very* early implementation of new pathfinding algorithm and logic. Uploaded as a backup. *UNTESTED*. Will probably not compile.

+++ gui/components/gui_convoy_assembler.cc
@@ -472,7 +472,7 @@ void gui_convoy_assembler_t::zeichnen(koord parent_pos)
 		}
 		max_speed = min(min_top_speed, (uint32) sqrt((((double)total_power/total_min_weight)-1)*2500));
 		min_speed = min(min_top_speed, (uint32) sqrt((((double)total_power/total_max_weight)-1)*2500));
-		uint8 tile_length;
+		uint16 tile_length;
 		if(depot_frame != NULL)
 		{
 			tile_length = depot_frame->get_convoy()->get_tile_length();
